Working depth buffer renderer
Renamed the source directory to src

Renamed raytracer.cpp/.h to main.cpp/.h

Creates math functions in math.cpp/.h to perform basic vector and matrix operations, but needs to be converted into a class and optimized.

Created primitives base class and a sphere class to extend upon it. Closes #4 and closes #3.

Created a world class to hold the primitives and update them. Going forward, it will get a scene graph and physics.

Created a ray tracer class that performs non-recursive ray casting to determine a depth buffer. For now, the depth buffer values are used to determine the color value for all color channels as % of depth between the min and max depth in the scene, scaled by 255 for each 8 bit channel.