Memoized primary rays
Added a destructor for the Camera class. Closes #27.

Modified the RayTracer class to precompute the primary rays to avoid overhead of recomputing them on each frame.  At higher resolutions, this can lead to speedups.  At 1024x768, this led to a 10% improvement in FPS from 34 to ~38.  It will become obsolete when moving to shaders. Closes #26.