Sped up ray-sphere intersection per Glassner, improved parallelism, and preparing for shader ray tracing
1) The ray-sphere intersection function was rewritten using a technique from "An Introduction to Ray Tracing" by Andrew S. Glassner.

The previous implementation first calculated the closest approach of the ray to the origin of the sphere as a vector projection and then took it's magnitude, resulting in one unnecessary square root and normalization per intersection. THe dot product would contain the magnitude of the vector projection, so the projection is skipped. The projection magnitude need not divide by the magnitude of the ray since it is normalized, avoiding the extra vector magnitude calculation.

Closes #32.

2) Vtune was used to identify if there were areas for overall optimization.  Pipeline stalls were prevalent. OpenMP was used to parallelize each pixel across a row, but not parallelize rows themselves. Swapping to parallelizing rows gave a slight performance improvement due to caching but pipeline stalls still occurred, presumably due to the sparse scene and some primary rays not intersecting with anything.

Another approach was taken to allow OpenMP to parallelize down rows and each pixel in a row, thereby enabling the thread to process the next pixel/primary ray, allowing filling of the pipeline when a non-intersecting primary ray is found.

Changing to parallelizing down rows and then across a row resulted in > 15% speedup in framerates but still has not drastically improved pipeline stalls.

3) Used the boilerplate shader loading/compiling/linking code from https://about-prog.com/opengl/ogl-triangle to prepare to migrate the ray tracer to GLSL shaders. Linkiing against Opengl32.lib.