Updated the scene and corrected buffer updates to the GPU
1) The scene was changed to include a spiral of spheres and an animated light.

2) The sphere ssbo that is used by the ray tracing compute shader was not being updated correctly.  Using glBufferSubData, the buffer as being modified by the CPU but a barrier wasn't preventing the CPU-side updates from occurring while the buffer was in use by the shader.

Instead of using glBufferSubData to modify the contents of the buffer, the new method first placed a barrier on the buffer and then maps the contents of the existing buffer to client space using glMapBuffer.  The contents are then modified and the buffer is unmapped.

Code was added to update the lights in render() and to update the contents of the buffer to the GPU.